SCC: Separating vectorisation test from general scc tests.
